common: image-android-dt: Provide means for id/rev-based DT lookup
The order in which DT blobs are passed to 'mkdtboimg.py create' [1] is
_not_ preserved in the binary contents of the image [2], assuming that
the latter is accurately represented by dumping the contents [3] of the
generated image (regardless if on the host or on the target). In other
words, when generating the DTB/DTBO image, users have apparently no way
to impose a certain order of DT blobs. Thus, any assumption that DT
entry X is located at index Y in the image is a dangerous assumption.
For the same reasons, trying to identify the desired DT blob simply
based on the index-based adtimg API is not suitable for the end user.

To correctly identify the desired DT entry, implement the field
(id/rev)-based DT retrieval. Along with the address and size of the
blob, also return the index of the identified DT, to fulfil below
requirement [3] (thanks to Sam for pointing it out):

 --- snip ---
 Bootloader must support the updated boot image and must add a new
 kernel command line parameter androidboot.dtb_idx to indicate the
 index of the selected DT to facilitate VTS verification. You can only
 specify one (1) index. For example, the parameter androidboot.dtb_idx=N
 reports N as the zero-based index of the device tree selected by the
 bootloader from the set of DT blobs present in the boot image.
 --- snip ---

This patch needs to be complemented by appropriate updates in the
frontend (adtimg).

[1] https://source.android.com/devices/architecture/dto/partitions#mkdtimg
[2] Verified using mkdtboimg.py version android-10.0.0_r20
[3] https://source.android.com/devices/architecture/dto/partitions#dump
[4] https://source.android.com/devices/bootloader/boot-image-header#boot-image-header-version-2

Cc: Sam Protsenko <semen.protsenko@linaro.org>
Signed-off-by: Eugeniu Rosca <erosca@de.adit-jv.com>